+++ CHANGELOG.md
@@ -6,7 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+
+## [0.16.1]
+
+### Added
+
+- `Attachment::with_content_id`, will replace `Attachment::with_inline_content_id` in the next
+  major release.
+
+## [0.16.0]
 
 ### Added
 
@@ -313,6 +322,8 @@ Disabled `reqwest`'s default features and enabled `rustls-tls`.
 
 Initial release.
 
+[0.16.1]: https://crates.io/crates/resend-rs/0.16.1
+[0.16.0]: https://crates.io/crates/resend-rs/0.16.0
 [0.15.0]: https://crates.io/crates/resend-rs/0.15.0
 [0.14.1]: https://crates.io/crates/resend-rs/0.14.1
 [0.14.0]: https://crates.io/crates/resend-rs/0.14.0